resolved merge conflicts. integrated sidebar to work with layout
